// LDC.Atlas.WebApp/Client/app/shared/dtos/accountingEntries-search-result.ts
export interface AccountingEntriesSearchResult {
    clientReference: string;
    costType: string;
    dmsId: number;
    department: string;
    associatedAccountCode: string;
    paymentTerm: string;
    physicalContractCode: string;
    contractSectionCode: string;
    valueDate: Date;
    weight: number;
    province: string;
    payee: string;
    roe: number;
    roeType: string;
    commodity: string;
    narrative: number;
    accountLineType: number;
    charter: string;
    postingLineId: number;
    accountingDocumentId: number;
    nomAccount: string;
    matchFlag: string;
    amount: number;
    currencyCode: string;
    accountingPeriod: string;
    documentDate: Date;
    postedDate: Date;
    functionalCurrency: number;
    statutoryCurrency: number;
    securePayment: string;
    vatTurnover: number;
    accountCategory: string;
    preMatch: string;
    interfaceStatus: string;
    originalReferenceId: string;
    accuralNumber: number;
    backOfficeDocId: string;
    glDate: Date;
    backOfficePostingDate: Date;
    documentType: string;
    postedTime: string;
    setupUser: string;
    setupDate: Date;
    setupTime: string;
    originalValueDate: Date;
    vatCode: string;
    matchDate: Date;
    matchTime: string;
    inhouseExternal: string;
    contractCommodityDescription: string;
    contractCommodityType: string;
    contractIncoTerms: string;
    contractBLDate: Date;
    clientGroupAccount: string;
    contractGroupNumber: number;
    contractPortOrigin: string;
    contractPortDestination: string;
    contractIntercomPort: string;
    contractShippingPeriod: string;
    contractTrader: string;
    contractVessel: string;
    contractVesselImo: string;
    accountingDate: Date;
    paymentDocumentDate: Date;
    secondaryReference: string;
    paymentTraxId: string;
    costCenter: string;
    clientAccount: string;
    sectionId: number;
    invoiceId: number;
    invoiceTypeId: number;
    cashId: number;
    costDirectionId: number;
    originalInvoiceTypeId: number;
    originalInvoiceId: number;
    transactionDocumentId: number;
    documentReference: string;
}